위키피디아 하이퍼링크 + 복수형 적용
